CrazyRandomSword class implemented and tested
CrazyRandomSword.h and CrazyRandomSword.cpp were created. When creating a new
weapon using the name "crazy", and sword is created where hitpoints = random
integer number between 7 and 100. Ignores a random amount of integer armor points, ranging from to 2 to a third of the armor the weapon hits.
The WeaponFactory.cpp file was adjusted to add the class and the main.cpp file
was editted in order to test the class.